pehchain-api/requirements.txt.txt
certifi (2017.11.5)
chardet (3.0.4)
falcon (1.3.0)
hug (2.3.2)
hug-middleware-cors (1.0.0)
idna (2.6)
olefile (0.44)
Pillow (4.3.0)
pip (9.0.1)
PyJWT (1.5.3)
pymongo (3.5.1)
pyotp (2.2.6)
python-mimeparse (1.6.0)
redis (2.10.6)
requests (2.18.4)
setuptools (38.2.3)
six (1.11.0)
urllib3 (1.22)
wheel (0.30.0)